requirements.txt
opencv-python == 4.11.0
numpy == 2.0.2
pillow == 11.2.1